requirements.txt
mcstatus>=2.2.1
discord.py>=1.1.1
toml>=0.10.0
